requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt setup.py
#
bison==0.1.0
bleach==3.1.0             # via readme-renderer
certifi==2019.3.9         # via requests
chardet==3.0.4            # via requests
docutils==0.14            # via readme-renderer
idna==2.8                 # via requests
jinja2==2.10.1
markupsafe==1.1.1         # via jinja2
pkginfo==1.5.0.1          # via twine
pygments==2.4.2           # via readme-renderer
pyyaml==5.1               # via bison
readme-renderer==24.0     # via twine
requests-toolbelt==0.9.1  # via twine
requests==2.22.0          # via requests-toolbelt, twine
six==1.12.0               # via bleach, readme-renderer
tqdm==4.32.1              # via twine
twine==1.13.0
urllib3==1.25.3           # via requests
webencodings==0.5.1       # via bleach
wheel==0.33.4
